updated Breweries table and changed seed file for lat and long values

+++ db/migrate/20210325224101_create_breweries.rb
@@ -2,15 +2,15 @@ class CreateBreweries < ActiveRecord::Migration[6.1]
   def change
     create_table :breweries do |t|
       t.string :name
-      t.integer :phone
+      t.bigint :phone
       t.string :brewery_type
       t.string :website_url
       t.string :street
       t.string :state
       t.string :city
-      t.integer :zip
-      t.integer :lat
-      t.integer :long
+      t.bigint :zip
+      t.numeric :lat
+      t.numeric :long
 
       t.timestamps
     end
